refactor(dashboard-rework): Moved the event query down
